+++ TDEngine2/plugins/WindowsInputContext/include/CWindowsInputContext.h
@@ -9,6 +9,7 @@
 
 #include <core/IInputContext.h>
 #include <core/CBaseObject.h>
+#include <core/Event.h>
 #include <utils/Utils.h>
 
 
@@ -69,7 +70,7 @@ namespace TDEngine2
 		on DirectInput and XInput APIs
 	*/
 
-	class CWindowsInputContext : public IDesktopInputContext, public CBaseObject
+	class CWindowsInputContext : public IDesktopInputContext, public CBaseObject, public IEventHandler
 	{
 		public:
 			friend TDE2_API IInputContext* CreateWindowsInputContext(TPtr<IWindowSystem>, E_RESULT_CODE&);
@@ -141,6 +142,25 @@ namespace TDEngine2
 			TDE2_API bool IsMouseButtonUnpressed(U8 button) override;
 
 			/*!
+				\brief The method receives a given event and processes it
+
+				\param[in] pEvent A pointer to event data
+
+				\return RC_OK if everything went ok, or some other code, which describes an error
+			*/
+
+			TDE2_API E_RESULT_CODE OnEvent(const TBaseEvent* pEvent) override;
+
+
+			/*!
+				\brief The method returns an identifier of a listener
+
+				\return The method returns an identifier of a listener
+			*/
+
+			TDE2_API TEventListenerId GetListenerId() const override;
+
+			/*!
 				\brief The method returns a position of a cursor
 
 				\return The method returns a position of a cursor
@@ -193,6 +213,10 @@ namespace TDEngine2
 			*/
 
 			TDE2_API TPtr<IGamepad> GetGamepad(U8 gamepadId) const override;
+
+#if TDE2_EDITORS_ENABLED
+			TDE2_API void SetOnCharInputCallback(const TOnCharActionCallback& onEventAction) override;
+#endif
 		protected:
 			DECLARE_INTERFACE_IMPL_PROTECTED_MEMBERS(CWindowsInputContext)
 
@@ -213,6 +237,10 @@ namespace TDEngine2
 			TPtr<IGamepad>     mpGamepads[mMaxNumOfGamepads];
 
 			TInternalInputData mInternalData;
+
+#if TDE2_EDITORS_ENABLED
+			TOnCharActionCallback mOnCharInputCallback;
+#endif
 	};
 }